feat: added manage error display in edit form
